# docs/swagger.yaml
definitions:
...
  dto.CheckPermissionDTO:
    properties:
      act:
        description: متد مانند GET, POST, PUT, DELETE
        type: string
      obj:
        description: مسیر مانند /user/all یا /approvals/:id
        type: string
      sub:
        description: می‌تواند نقش یا یوزر باشد
        type: string
    required:
    - act
    - obj
    - sub
    type: object
...
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      full_name:
        maxLength: 100
        type: string
      mobile:
        type: string
      national_code:
        type: string
      password:
        maxLength: 100
        minLength: 6
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - mobile
    - password
    - username
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      mobile:
        type: string
      uid:
        type: string
      username:
        type: string
    type: object
...
  response.Meta:
    properties:
      code:
        type: integer
      extra:
        additionalProperties: true
        description: For additional metadata
        type: object
      message:
        type: string
      message_id:
        description: For i18n
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      success:
        type: boolean
      trace_id:
        type: string
    type: object
  response.PaginationMeta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      meta:
        $ref: '#/definitions/response.Meta'
    type: object
info:
  contact: {}
  description: This is a sample API for demonstrating Swagger with Bearer Authentication
    in Go using Gin
  title: Swagger Example API
  version: "1.0"
paths:
  /acl/check:
    get:
      consumes:
      - application/json
      description: بررسی می‌کند که آیا کاربر اجازه انجام عمل مشخصی را دارد یا خیر.
      parameters:
      - description: اطلاعات بررسی دسترسی
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CheckPermissionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: بررسی سطح دسترسی
      tags:
      - ACL
...
  /acl/permission/list:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      tags:
      - ACL
  /acl/policy/create:
    post:
      consumes:
      - application/json
      description: مجوزی برای یک کاربر/گروه جهت انجام عملی بر یک شیء اضافه می‌کند.
      parameters:
      - description: اطلاعات مجوز
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CheckPermissionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: افزودن مجوز
      tags:
      - ACL
  /acl/policy/remove:
    delete:
      consumes:
      - application/json
      description: مجوز مشخص شده را برای کاربر/گروه حذف می‌کند.
      parameters:
      - description: اطلاعات مجوز برای حذف
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CheckPermissionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: حذف مجوز
      tags:
      - ACL
...
  /routes/list:
    get:
      consumes:
      - application/json
      description: این متد تمام مسیرها را همراه با متد HTTP آن‌ها دسته‌بندی‌شده بر
        اساس prefix بازمی‌گرداند.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: دریافت لیست اندپوینت‌ها
      tags:
      - Route
  /user/{uid}:
    delete:
      description: Delete user by UID (admin only)
      parameters:
      - description: User UID
        in: path
        name: uid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user details by UID (admin or self)
      parameters:
      - description: User UID
        in: path
        name: uid
        required: true
        type: string
      - description: User update data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a user
      tags:
      - users
  /user/all:
    get:
      description: Retrieve a list of all users
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all users
      tags:
      - users
  /user/create:
    post:
      consumes:
      - application/json
      description: Register a new user in the system
      parameters:
      - description: User data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new user
      tags:
      - users
  /user/me:
    get:
      description: Get authenticated user information from token
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get current user info
      tags:
      - users
  /user/uid/{uid}:
    get:
      description: Get a user by their UID
      parameters:
      - description: User UID
        in: path
        name: uid
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Find user by UID
      tags:
      - users
  /user/username/{username}:
    get:
      description: Get a user object by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Find user by username
      tags:
      - users
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"